Fixed image display
Image was a double and needed to be cast to an int before being converted into an image.

+++ CNN1/Form1.cs
@@ -129,7 +129,7 @@ namespace CNN1
                     {
                         for (int ii = 0; ii < scale; ii++)
                         {
-                            scaled[(j * scale) + i, (jj * scale) + ii] = image[jj, j];
+                            scaled[(j * scale) + i, (jj * scale) + ii] = (int)image[jj, j];
                         }
                     }
                 }
